[#98] Move common filters to lib
