+++ iceoryx2-cli/iox2-service/src/cli.rs
@@ -13,8 +13,8 @@
 use clap::Args;
 use clap::Parser;
 use clap::Subcommand;
-use clap::ValueEnum;
 
+use iceoryx2_cli::filter::MessagingPatternFilter;
 use iceoryx2_cli::help_template;
 use iceoryx2_cli::Format;
 
@@ -34,16 +34,6 @@ pub struct Cli {
 
     #[clap(long, short = 'f', value_enum, global = true, value_enum, default_value_t = Format::Ron)]
     pub format: Format,
-}
-
-#[derive(Debug, Clone, ValueEnum)]
-#[clap(rename_all = "PascalCase")]
-#[derive(Default)]
-pub enum MessagingPatternFilter {
-    PublishSubscribe,
-    Event,
-    #[default]
-    All,
 }
 
 #[derive(Debug, Clone, Args)]
